python-analytics/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.5.0
statsmodels>=0.14.0
ta>=0.11.0
yfinance>=0.2.40
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
